+++ jdaviz/configs/cubeviz/plugins/moment_maps/moment_maps.py
@@ -29,8 +29,7 @@ moment_unit_options = {0: ["Surface Brightness"],
                        2: ["Velocity", "Velocity^N"]}
 
 
-@tray_registry('cubeviz-moment-maps', label="Moment Maps",
-               viewer_requirements=['spectrum', 'image'])
+@tray_registry('cubeviz-moment-maps', label="Moment Maps")
 class MomentMap(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixin,
                 SpectralContinuumMixin, AddResultsMixin):
     """