+++ jdaviz/configs/default/plugins/subset_tools/tests/test_subset_tools.py
@@ -6,7 +6,8 @@ import pytest
 from astropy import units as u
 from astropy.coordinates import SkyCoord
 from astropy.nddata import NDData
-from glue.core.edit_subset_mode import ReplaceMode, OrMode, NewMode
+from astropy.tests.helper import assert_quantity_allclose
+from glue.core.edit_subset_mode import ReplaceMode, NewMode
 from glue.core.roi import EllipticalROI, CircularROI, CircularAnnulusROI, RectangularROI
 from numpy.testing import assert_allclose
 from regions import (CircleAnnulusPixelRegion, CirclePixelRegion,
@@ -111,6 +112,8 @@ def test_circle_recenter_linking(roi_class, subset_info, imviz_helper, image_2d_
     plugin = imviz_helper.plugins['Subset Tools']._obj
     assert not plugin.display_sky_coordinates
 
+    plugin.subset.selected = 'Subset 1'
+
     # get initial subset definitions from ROI applied
     subset_defs = plugin.subset_definitions
 
@@ -153,6 +156,8 @@ def test_circle_recenter_linking(roi_class, subset_info, imviz_helper, image_2d_
     new_pix_region = original_sky_region.to_pixel(img_wcs)
     new_roi = regions2roi(new_pix_region)
     imviz_helper.plugins['Subset Tools'].import_region(new_roi)
+
+    plugin.subset.selected = 'Subset 2'
 
     # get subset definitions again, which should now be in sky coordinates
     subset_defs = plugin.subset_definitions
@@ -218,8 +223,8 @@ def test_import_spectral_regions_file(cubeviz_helper, spectrum1d_cube, tmp_path)
     subsets = cubeviz_helper.app.get_subsets()
     assert len(subsets) == 1
 
-    plg.combination_mode = 'or'
-    plg.import_region(SpectralRegion(7 * u.um, 8 * u.um))
+    plg.import_region(SpectralRegion(7 * u.um, 8 * u.um), edit_subset='Subset 1',
+                      combination_mode='or')
 
     subsets = cubeviz_helper.app.get_subsets()
     assert len(subsets['Subset 1']) == 2
@@ -228,7 +233,7 @@ def test_import_spectral_regions_file(cubeviz_helper, spectrum1d_cube, tmp_path)
                SpectralRegion(5.8 * u.um, 5.9 * u.um))
     plg.import_region(subset2, combination_mode=['new', 'andnot'])
 
-    assert cubeviz_helper.app.session.edit_subset_mode.mode == OrMode
+    assert cubeviz_helper.app.session.edit_subset_mode.mode == ReplaceMode
 
     with pytest.raises(ValueError, match='\'test\' not one of'):
         plg.combination_mode = 'test'
@@ -372,8 +377,8 @@ def test_get_regions_composite_wcs_linked(imviz_helper, image_2d_wcs):
                           dec=-20.808486*u.deg), radius=0.008*u.deg)
     sr2 = CircleSkyRegion(center=SkyCoord(ra=337.51*u.deg, dec=-20.81*u.deg),
                           radius=0.007*u.deg)
-    st.import_region(sr1, combination_mode='new')
-    st.import_region(sr2, combination_mode='and')
+    st.import_region(sr1)
+    st.import_region(sr2, edit_subset='Subset 1', combination_mode='and')
 
     # composite subset should be a sky region, and combined to a compound region
     regs = st.get_regions(return_sky_region=True)
@@ -385,11 +390,10 @@ def test_get_regions_composite_wcs_linked(imviz_helper, image_2d_wcs):
     assert cr.operator == operator.and_
 
 
-@pytest.mark.skip(reason="Unskip after JDAT-5186.")
 def test_get_composite_sky_region_remove(imviz_helper, image_2d_wcs):
     """
-    Test to ensure bug fixed by JDAT-5186 is fixed, where get_subsets
-    for compoisute subset applied with 'remove' when WCS linked was not
+    Test to ensure bug is fixed, where get_subsets
+    for composite subset applied with 'remove' when WCS linked was not
     correctly retrieving the second subset.
     """
     data = NDData(np.ones((128, 128)) * u.nJy, wcs=image_2d_wcs)
@@ -403,21 +407,24 @@ def test_get_composite_sky_region_remove(imviz_helper, image_2d_wcs):
                           dec=-20.808486*u.deg), radius=0.008*u.deg)
     sr2 = CircleSkyRegion(center=SkyCoord(ra=337.51*u.deg, dec=-20.81*u.deg),
                           radius=0.007*u.deg)
-    st.import_region(sr1, combination_mode='new')
-    st.import_region(sr2, combination_mode='andnot')
+    regions = [sr1, sr2]
+    st.import_region(regions, combination_mode=['new', 'andnot'])
 
-    # app.get_subsets
+    # app.get_subsets, composite region should be returned as 2 sky regions
     ss = imviz_helper.app.get_subsets(include_sky_region=True)
-
-    # composite region should be returned as 2 sky regions
-    assert isinstance(ss['Subset 1'][0]['sky_region'], CircleSkyRegion)
-    assert isinstance(ss['Subset 1'][1]['sky_region'], CircleSkyRegion)
+    for expected_region, actual in zip(regions, ss['Subset 1']):
+        sky_region = actual['sky_region']
+        assert_quantity_allclose(sky_region.center.ra, expected_region.center.ra)
+        assert_quantity_allclose(sky_region.center.dec, expected_region.center.dec)
+        assert_quantity_allclose(sky_region.radius, expected_region.radius)
 
     # now make sure Subset Tools get_regions agrees and both sky regions
     # are returned for Subset 1
     regs = st.get_regions()
-    assert isinstance(regs['Subset 1'][0], CircleSkyRegion)
-    assert isinstance(regs['Subset 1'][1], CircleSkyRegion)
+    for expected_region, actual in zip(regions, regs['Subset 1']):
+        assert_quantity_allclose(actual[0].center.ra, expected_region.center.ra)
+        assert_quantity_allclose(actual[0].center.dec, expected_region.center.dec)
+        assert_quantity_allclose(actual[0].radius, expected_region.radius)
 
 
 def test_check_valid_subset_label(imviz_helper):
@@ -430,10 +437,10 @@ def test_check_valid_subset_label(imviz_helper):
 
     # apply three subsets, with their default names of `Subset 1`, `Subset 2`, and `Subset 3`
     st.import_region(CircularROI(20, 20, 10))
-    st.subset = "Create New"
     st.import_region(CircularROI(25, 25, 10))
-    st.subset = "Create New"
     st.import_region(CircularROI(30, 30, 10))
+
+    st.subset.selected = "Subset 3"
 
     # we should not be able to rename or add a subset named 'Subset 1'.
     # Make sure this warns and returns accordingly.
@@ -450,15 +457,16 @@ def test_rename_subset(cubeviz_helper, spectrum1d_cube):
     plg = cubeviz_helper.plugins['Subset Tools']
 
     spatial_reg = CirclePixelRegion(center=PixCoord(x=2, y=2), radius=2)
-    plg.import_region(spatial_reg, combination_mode='new')
+    plg.import_region(spatial_reg)
     spatial_reg = CirclePixelRegion(center=PixCoord(x=4, y=4), radius=1)
-    plg.import_region(spatial_reg, combination_mode='new')
+    plg.import_region(spatial_reg)
 
     plg.rename_subset("Subset 1", "Test Rename")
 
     assert plg.subset.choices == ['Create New', 'Test Rename', 'Subset 2']
     assert cubeviz_helper.app.data_collection[-1].label == "Spectrum (Test Rename, sum)"
 
+    plg.subset = "Subset 2"
     plg.rename_selected("Second Test")
     assert plg.subset.choices == ['Create New', 'Test Rename', 'Second Test']
 
@@ -471,9 +479,9 @@ def test_update_subset(cubeviz_helper, spectrum1d_cube):
     plg = cubeviz_helper.plugins['Subset Tools']
 
     spatial_reg = CirclePixelRegion(center=PixCoord(x=2, y=2), radius=2)
-    plg.import_region(spatial_reg, combination_mode='new')
+    plg.import_region(spatial_reg)
     spatial_reg = CirclePixelRegion(center=PixCoord(x=4, y=4), radius=1)
-    plg.import_region(spatial_reg, combination_mode='and')
+    plg.import_region(spatial_reg, edit_subset='Subset 1', combination_mode='and')
 
     subset_def = plg.update_subset('Subset 1')
     assert isinstance(subset_def, dict)